corrige dags 04 e 05
